+++ api/api.go
@@ -22,15 +22,16 @@ func SetConfig(cfg *cfg.Config) {
 // Commands is the main function of this package
 func Commands() *mux.Router {
 	rtr := mux.NewRouter()
-	rtr.HandleFunc("/v0/container/{taskID}", V0StatusContainer).Methods("GET")
 	rtr.HandleFunc("/v0/agent/scale/{count}", V0ScaleK3SAgent).Methods("GET")
-	rtr.HandleFunc("/v0/agent/reflate", V0ReflateK3SAgent).Methods("GET")
+	rtr.HandleFunc("/v0/agent/scale", V0GetCountK3SAgent).Methods("GET")
 	rtr.HandleFunc("/v0/server/scale/{count}", V0ScaleK3S).Methods("GET")
-	rtr.HandleFunc("/v0/server/reflate", V0ReflateK3S).Methods("GET")
+	rtr.HandleFunc("/v0/server/scale", V0GetCountK3S).Methods("GET")
 	rtr.HandleFunc("/v0/server/config", V0GetKubeconfig).Methods("GET")
+	rtr.HandleFunc("/v0/server/version", V0GetKubeVersion).Methods("GET")
 	rtr.HandleFunc("/v0/etcd/scale/{count}", V0ScaleEtcd).Methods("GET")
 	rtr.HandleFunc("/v0/task/kill/{id}", V0KillTask).Methods("GET")
 	rtr.HandleFunc("/v0/status/m3s", V0StatusM3s).Methods("GET")
+	rtr.HandleFunc("/v0/status/k8s", V0StatusK8s).Methods("GET")
 
 	return rtr
 }