+++ mesos_cli/m3s/main.py
@@ -135,7 +135,7 @@ class M3s(PluginBase):
             ) from exception
 
         framework_address = get_framework_address(
-            argv["<framework-id>"], master, config
+            self.get_framework_id(argv), master, config
         )
         data = http.read_endpoint(framework_address, "/v0/server/config", self)
 
@@ -157,7 +157,7 @@ class M3s(PluginBase):
             ) from exception
 
         framework_address = get_framework_address(
-            argv["<framework-id>"], master, config
+            self.get_framework_id(argv), master, config
         )
         data = http.read_endpoint(framework_address, "/v0/server/version", self)
 
@@ -179,13 +179,16 @@ class M3s(PluginBase):
             ) from exception
 
         framework_address = get_framework_address(
-            argv["<framework-id>"], master, config
+            self.get_framework_id(argv), master, config
         )
 
         if argv["--m3s"]:
             data = http.read_endpoint(framework_address, "/v0/status/m3s", self)
+            print(data)
 
-        print(data)
+        if argv["--kubernetes"]:
+            data = http.read_endpoint(framework_address, "/v0/status/k8", self)
+            print(data)
 
     def list(self, argv):
         """
@@ -225,16 +228,34 @@ class M3s(PluginBase):
 
         print(str(table))
 
+    def get_framework_id(self, argv):
+        """
+        Resolv the id of a framework by the name of a framework
+        """
+
+        if argv["<framework-id>"].count("-") != 5:
+            data = get_frameworks(self.config.master(), self.config)
+            for framework in data:
+                if (
+                    framework["active"] is not True
+                    or framework["name"].lower() != argv["<framework-id>"].lower()
+                ):
+                    continue
+                return framework["id"]
+        return argv["<framework-id>"]
+
     def principal(self):
         """
         Return the principal in the configuration file
         """
+
         return self.m3sconfig["m3s"].get("principal")
 
     def secret(self):
         """
         Return the secret in the configuration file
         """
+
         return self.m3sconfig["m3s"].get("secret")
 
     # pylint: disable=no-self-use
@@ -242,6 +263,7 @@ class M3s(PluginBase):
         """
         Return the connection timeout of the agent
         """
+
         return default
 
     def _get_config(self):